adicionado a variavel int no lugar de double

+++ java.java
@@ -2,7 +2,7 @@ class TestaCondicional {
     public static void main(String[] args) {
         System.out.println("Ola peopel");
 
-        double valor = 10;
+        int valor = 10;
 
         if (valor > 1 && valor < 5) {
             System.out.println("vc tem entre 1 e 5");
